requirements.txt
Django==5.1.2
psycopg[binary]==3.2.3
django-environ==0.11.2
whitenoise==6.7.0
django-allauth==0.62.1
requests>=2.31,<3
PyJWT>=2,<3
cryptography>=41
google-api-python-client>=2.131.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1"
